requirements.txt
selenium
chromedriver.exe
os
python-dotenv